Added exceptions

+++ src/main/java/com/uzairproject/librarymanagement/services/serviceimpl/AuthorServiceImpl.java
@@ -3,6 +3,9 @@ package com.uzairproject.librarymanagement.services.serviceimpl;
 import com.uzairproject.librarymanagement.dtos.AuthorRequestDto;
 import com.uzairproject.librarymanagement.dtos.AuthorResponseDto;
 import com.uzairproject.librarymanagement.dtos.BookResponseDto;
+import com.uzairproject.librarymanagement.exceptions.AuthorHasBooksException;
+import com.uzairproject.librarymanagement.exceptions.AuthorNotFoundException;
+import com.uzairproject.librarymanagement.exceptions.EmailAlreadyUsedException;
 import com.uzairproject.librarymanagement.models.Author;
 import com.uzairproject.librarymanagement.models.Book;
 import com.uzairproject.librarymanagement.repositories.AuthorRepo;
@@ -29,6 +32,10 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public AuthorResponseDto createAuthor(AuthorRequestDto authorDto){
+        Author existingAuthor = authorRepo.findByEmail(authorDto.email());
+        if (existingAuthor != null){
+            throw new EmailAlreadyUsedException("Email is already in use by another author");
+        }
         Author author = entityMappers.convertAuthorDtoToAuthorEntity(authorDto);
         Author savedAuthor = authorRepo.save(author);
         return entityMappers.convertAuthorEntityToAuthorDto(savedAuthor);
@@ -43,13 +50,13 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public AuthorResponseDto getAuthorById(Long authorId) {
-        Author author = authorRepo.findById(authorId).orElseThrow(() -> new RuntimeException("Author Not found: " + authorId));
+        Author author = authorRepo.findById(authorId).orElseThrow(() -> new AuthorNotFoundException("Author Not found: " + authorId));
         return entityMappers.convertAuthorEntityToAuthorDto(author);
     }
 
     @Override
     public AuthorResponseDto updateAuthorInfo(Long authorId, AuthorRequestDto authorDto) {
-        Author existingAuthor = authorRepo.findById(authorId).orElseThrow(() -> new RuntimeException("Author Not found: " + authorId));
+        Author existingAuthor = authorRepo.findById(authorId).orElseThrow(() -> new AuthorNotFoundException("Author Not found: " + authorId));
         Author updatedAuthor = entityMappers.convertAuthorDtoToAuthorEntity(authorDto);
         existingAuthor.setName(updatedAuthor.getName());
         existingAuthor.setEmail(updatedAuthor.getEmail());
@@ -60,8 +67,14 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public void deleteAuthor(Long authorId) {
-        Author author = authorRepo.findById(authorId).orElseThrow(() -> new RuntimeException("Author Not found: " + authorId));
-        authorRepo.delete(author);
+        List<Book> books = bookRepo.findByAuthorId(authorId);
+        if (books.isEmpty()){
+            Author author = authorRepo.findById(authorId).orElseThrow(() -> new AuthorNotFoundException("Author Not found: " + authorId));
+            authorRepo.delete(author);
+        }
+        else {
+            throw new AuthorHasBooksException("Before deleting author remove their books.");
+        }
     }
 
     @Override
